split account page into more components
